Centralize RandRange prototype into game.h (conform mis-typed definition)
The 7 callers drifted across s32/u32/u8 returns and s32/u8/s16 args. Disassembly
shows the function masks range->u8 at entry and the result->u8 at exit (the body's
(u8) cast), but returns the masked value for callers to use directly. The true
interface is s32 RandRange(u8 range); the definition was mis-typed u8-return.
Corrected the definition's return type (byte-neutral: the body cast still emits
the exit mask) -> one canonical prototype now serves all 7 callers with NO
call-site casts. Diagnosed the mis-decompiled side from the binary rather than
forcing a match with casts. Clean-build verified (make tidy && make && make check).

+++ src/system/init.c
@@ -25,7 +25,10 @@ void ClearVramAndWorkram(void)
 /* Step the LCG seed and return the result modulo `range`.
  * Multiplier 109 / increment 1021 — small-period generator used for
  * gameplay randomness (callers TBD; verified via struct_xref on offset 28). */
-u8 RandRange(u8 range)
+/* range is a byte (the binary masks it at entry); the result is narrowed to a
+ * byte by the (u8) cast in the body, but the interface returns it widened so
+ * callers use the masked value directly without re-masking. */
+s32 RandRange(u8 range)
 {
     gGameStuff.rngSeed = gGameStuff.rngSeed * 109 + 1021;
     return (u8)(gGameStuff.rngSeed % range);
